adding a Modeler warning for ObjAttributes mapped to generated PKs - we can't handle this well in the framework.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/cayenne/main/trunk@404511 13f79535-47bb-0310-9956-ffa450edef68

+++ cayenne/cayenne-java/src/tests/java/org/objectstyle/cayenne/access/IdentityColumnsTst.java
@@ -327,9 +327,15 @@ public class IdentityColumnsTst extends CayenneTestCase {
         MeaningfulGeneratedColumnTest o = (MeaningfulGeneratedColumnTest) context
                 .createAndRegisterNewObject(MeaningfulGeneratedColumnTest.class);
         o.setName("o1");
-        o.setGeneratedColumn(new Integer(2));
+        o.setGeneratedColumn(new Integer(33333));
 
         context.commitChanges();
         assertNotNull(o.getGeneratedColumn());
+
+        // Note - this only *appears* to work, but the following assertion will fail. For
+        // now we address the issue by giving a warning in the modeler that generated PKs
+        // can't be meaningful.
+        // assertEquals(new Integer(33333), o.getObjectId().getIdSnapshot().get(
+        // MeaningfulGeneratedColumnTest.GENERATED_COLUMN_PK_COLUMN));
     }
 }
